requirements.txt
requests
Pillow
opencv-python-headless
numpy
matplotlib
jsonlines
tqdm